# openapi.json
{
  "components": {
...
    "securitySchemes": {
      "access-token": {
        "bearerFormat": "JWT",
        "in": "header",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Gym Marketplace & Multi-Tenant Gym Management SaaS.\n\nErrors use the §C3.1 envelope: `{ error: { code, message, details[], correlation_id } }`.\n\nMoney crosses the wire as a STRING of integer minor units (paise for INR), never a JSON number — a JSON number loses integer precision above 2^53 (TR-38, BR-PAY-01).\n\nThe tenant is derived from the access token and is NEVER accepted from a client header, path, query or body (BR-TEN-01, §11.3).",
    "title": "GymMap API",
    "version": "1"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/auth/sessions/{sessionId}": {
      "delete": {
        "description": "A session belonging to anyone else is a 404 and not a 403 — a 403 confirms the uuid exists, and \"not yours\" and \"already revoked\" must be one answer. Revocation reaches the already-issued access token through the family denylist (AC-10) rather than waiting for the next refresh to fail.",
        "operationId": "revokeSession",
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Revokes one of the calling member’s own sessions.",
        "tags": [
          "iam"
        ],
        "x-gymmap-error-codes": [
          "UNAUTHENTICATED",
          "PERMISSION_DENIED",
          "RESOURCE_NOT_FOUND"
        ],
        "x-gymmap-permission": "iam.own_session.revoke",
        "x-gymmap-rate-limit": "RL-WRITE"
      }
    },
...
  },
  "servers": [],
...
}